+++ Loader/src/ModActions.cs
@@ -49,33 +49,25 @@ namespace Sheepy.Modnix {
          foreach ( var act in actions ) {
             foreach ( var dll in ActionHandlers ) {
                var result = RunActionHandler( mod, dll, act );
-               if ( result is Exception ex && AbortOnError( log, act, ex ) )
-                  return;
+               if ( result is Exception ex ) return;
             }
          }
          foreach ( var dll in ActionHandlers )
             ActionMods[ dll ].Log().Filters.Remove( modPrefix );
       } catch ( Exception ex ) { mod.Log().Error( ex ); } }
 
-      internal static ActionDef[] FilterActions ( ActionDef[] actions, string phase ) {
+      private static ActionDef[] FilterActions ( ActionDef[] actions, string phase ) {
          phase = phase.ToLowerInvariant();
          var result = actions.Where( e => InList( e["phase"]?.ToString(), phase ) ).ToArray();
          return result.Length > 0 ? result : null;
       }
 
-      internal static bool AbortOnError ( Logger log, ActionDef act, Exception err ) {
-         act.TryGetValue( "onerror", out object onerror );
-         var handle = TrimAndLower( onerror ) ?? "log";
-         if ( handle.IndexOf( "log" ) >= 0 ) log.Error( err );
-         else if ( handle.IndexOf( "err" ) >= 0 ) log.Error( err );
-         else if ( handle.IndexOf( "warn" ) >= 0 ) log.Warn( err );
-         else if ( handle.IndexOf( "info" ) >= 0 ) log.Info( err );
-         else if ( handle.IndexOf( "verbo" ) >= 0 ) log.Verbo( err );
-         if ( handle.IndexOf( "stop" ) >= 0 ) {
-            log.Info( "Aborting because OnError == Stop ({0})", handle );
-            return true;
-         }
-         return false;
+      private static void LogActionError ( Logger log, ActionDef act, Exception err ) {
+         var directives = act.GetText( "onerror", "log" );
+         if ( InList( directives, "log" ) || InList( directives, "error" ) ) log.Error( err );
+         else if ( InList( directives, "warn" ) ) log.Warn( err );
+         else if ( InList( directives, "info" ) ) log.Info( err );
+         else if ( InList( directives, "verbo" ) ) log.Verbo( err );
       }
 
       internal static HashSet< string > FindPhases ( ActionDef[] actions ) {
@@ -96,14 +88,15 @@ namespace Sheepy.Modnix {
          return found.Count > 0 ? found : null;
       }
 
-      private static Dictionary< string, HashSet< string > > StrLists = new Dictionary<string, HashSet<string>>();
+      private static readonly Dictionary< string, HashSet< string > > StrLists = new Dictionary<string, HashSet<string>>();
 
-      internal static bool InList ( string list, string val ) {
+      private static bool InList ( string list, string val ) {
          if ( string.IsNullOrWhiteSpace( list ) ) return false;
+         list = list.Trim().ToLowerInvariant();
          HashSet<string> parsed;
          lock ( StrLists ) {
             if ( ! StrLists.TryGetValue( list, out parsed ) ) {
-               parsed = new HashSet<string>( list.ToLowerInvariant().Split( new char[]{ ',' }, StringSplitOptions.RemoveEmptyEntries ).Select( e => e.Trim() ) );
+               parsed = new HashSet<string>( list.Split( new char[]{ ',' }, StringSplitOptions.RemoveEmptyEntries ).Select( e => e.Trim() ) );
                if ( parsed.Count == 0 ) parsed = null;
                StrLists.Add( list, parsed );
             }
@@ -115,7 +108,7 @@ namespace Sheepy.Modnix {
          ActionDef defValues = null;
          var actions = new List<ActionDef>();
          foreach ( var a in list ) {
-            if ( a.ContainsKey( "action" ) && string.Equals( a[ "action" ]?.ToString(), "default", StringComparison.InvariantCultureIgnoreCase ) )
+            if ( string.Equals( a.GetText( "action" ), "default", StringComparison.InvariantCultureIgnoreCase ) )
                MergeDefAction( ref defValues, a );
             else
                actions.Add( AddDefAction( a, defValues ) );
@@ -123,7 +116,7 @@ namespace Sheepy.Modnix {
          return actions.Count > 0 ? actions.ToArray() : null;
       }
 
-      internal static object RunActionHandler ( ModEntry mod, DllMeta dll, ActionDef act ) { try {
+      private static object RunActionHandler ( ModEntry mod, DllMeta dll, ActionDef act ) { try {
          var lib = ModPhases.LoadDll( mod, dll.Path );
          if ( lib == null ) return false;
          var handler = ActionMods[ dll ];
@@ -135,14 +128,18 @@ namespace Sheepy.Modnix {
             if ( result is bool success ) {
                if ( success ) return true;
             } else if ( result is Exception ex ) {
+               LogActionError( handler.Log(), act, ex );
+               if ( InList( act.GetText( "onerror" ), "continue" ) ) continue;
+               if ( InList( act.GetText( "onerror" ), "skip" ) ) return null;
+               mod.Log().Info( "Aborting Actions. Set OnError to Continue or Skip to not abort." );
                return ex;
             } else if ( result != null )
-               handler.Log().Info( "Unexpected ActionMod result: {0}", result.GetType() );
+               handler.Log().Error( "Unexpected ActionMod result: {0}", result.GetType() );
          }
          return null;
       } catch ( Exception ex ) { mod.Error( ex ); return null; } }
 
-      internal static object ParamValue ( ActionDef act, ParameterInfo arg, ModEntry actionMod, ModEntry handler ) {
+      private static object ParamValue ( ActionDef act, ParameterInfo arg, ModEntry actionMod, ModEntry handler ) {
          if ( arg.ParameterType == typeof( ActionDef ) )
             return act;
          if ( arg.ParameterType == typeof( string ) )
@@ -150,7 +147,10 @@ namespace Sheepy.Modnix {
          return ModPhases.ParamValue( arg, handler );
       }
 
-      private static string TrimAndLower ( object key ) => key?.ToString().Trim().ToLowerInvariant();
+      private static string GetText ( this ActionDef act, string key, string fallback = null ) {
+         if ( act.TryGetValue( key, out object o ) && o is string txt ) return txt;
+         return fallback;
+      }
 
       private static void MergeDefAction ( ref ActionDef defValues, ActionDef a ) {
          if ( defValues == null ) defValues = new ActionDef();